first entry

+++ internal/storage/storage.go
@@ -1,6 +1,6 @@
 package storage
 
 type Storage interface{
-	CreateStudent(name string,email string,age int)(int,error)
+	CreateStudent(name string,email string,age int)(int64,error)
 }
